Creating and Joining a server works now but..
A server is created if there isn't one and then the server is joined if
one already exists.  The controls don't control other planes.  However
data being passed isn't good enough as you can see when there is more
than one plane in a server they appear to look as if they are buggy from
the other person's game view